DRYing public types
