+++ src/components/Point/Point2d/point2d-plugin.ts
@@ -1,17 +1,12 @@
 import { getPointSchema, sanitizePoint, formatPoint, normalizePoint, Format } from '../point-plugin'
-import { NumberSettings, InternalNumberSettings } from '../../Number/number-plugin'
+import { InternalNumberSettings } from '../../Number/number-plugin'
 import { InputWithSettings } from '../../../types/'
+import { Point2d, Point2dObject, Point2dSettings } from '../../../types/public-api-types'
 
-type Point2dArray = [number, number, number]
-type Point2dObj = { x: number; y: number }
-export type Point2d = Point2dObj | Point2dArray
-
-type Point2dSettings = { [key in keyof Point2dObj]?: NumberSettings }
-export type InternalPoint2dSettings = { [key in keyof Point2dObj]: InternalNumberSettings } & { format: Format }
-
+export type InternalPoint2dSettings = { [key in keyof Point2dObject]: InternalNumberSettings } & { format: Format }
 type Point2dInput = InputWithSettings<Point2d, Point2dSettings>
 
-export const KEYS: (keyof Point2dObj)[] = ['x', 'y']
+export const KEYS: (keyof Point2dObject)[] = ['x', 'y']
 export const schema = getPointSchema(KEYS)
 
 export const sanitize = (v: any, settings: InternalPoint2dSettings) => sanitizePoint(v, settings, KEYS) as Point2d
