+++ src/components/Point/Point3d/Point3d.tsx
@@ -1,9 +1,10 @@
 import React from 'react'
 import styled from '@xstyled/styled-components'
 import { LevaInputProps } from '../../../types/'
+import { Point3d as Point3dType } from '../../../types/public-api-types'
 import { PointCoordinates } from '../../PointCoordinates'
 import { Row, Label } from '../../styles'
-import { Point3d as Point3dType, KEYS, InternalPoint3dSettings } from './point3d-plugin'
+import { KEYS, InternalPoint3dSettings } from './point3d-plugin'
 import { mapArrayToKeys } from '../../../utils'
 
 type Point3dProps = LevaInputProps<Point3dType, InternalPoint3dSettings>
